Fixes #1
Though I agree with some of the formatting changes, the large majority goes against the standards I subscribe to. In particular:
- YAML supports one or more spaces, I prefer two for horizontal readability.
- Sass/SCSS/CSS does the same, and I prefer the same.
- In Twig/HTML I prefer four spaces, but this is a refactor that will coincide with an overhaul of the base code. A lot is inherited from from the Masonry-theme.
- In PHP my setup is more contextual to the logic the file, function, class, and methods, but the standards I follow are fairly strict usually. Admittedly I need to re-synchronize the CodeSniffer-setups between my development environments.

Since Git does not allow us to stage hunks or lines from commits, I am re-committing the changes applicable to be able to do it in a granular fashion.